pyproject.toml
[tool.poetry]
name = "sistema-alfa-montagem"
version = "0.1.0"
description = ""
authors = ["jeffers <jeffersonlima5286@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.3"
django-rest-framework = "^0.1.0"
djangorestframework-simplejwt = "^5.5.0"
python-decouple = "^3.8"
cloudinary = "^1.44.1"
django-cloudinary-storage = "^0.3.0"
pillow = "^11.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
